refactor: Organize project structure
- Move test files to tests/ folder
- Move CLI scripts to scripts/ folder
- Move data files to data/ folder
- Move core services to src/services/
- Move config.js to config/index.js
- Update all import paths throughout codebase
- Clean root directory (only index.js + config folders remain)

Project structure is now:
- config/       - Configuration files
- dashboard/    - React dashboard
- data/         - Runtime data files
- docs/         - Documentation
- migrations/   - DB migrations
- routes/       - Express routes
- scripts/      - CLI tools and utilities
- src/          - Source code (services, agents, commands, utils)
- tests/        - Test files
