+++ scripts/agent-cli.js
