+++ scripts/deep-dive-discovery.js
