+++ scripts/diagnose-issues.js
