+++ src/services/discord-handlers-parts/part-00.js
@@ -17,37 +17,37 @@ const {
     AutoModerationRuleTriggerType,
     DiscordAPIError
 } = require('discord.js');
-const JarvisAI = require('./jarvis-core');
-const config = require('./config');
-const braveSearch = require('./brave-search');
+const JarvisAI = require('../jarvis-core');
+const config = require('../../../config');
+const braveSearch = require('../brave-search');
 const { createCanvas, loadImage, registerFont } = require('canvas');
 const sharp = require('sharp');
 const fs = require('fs');
 const path = require('path');
-const database = require('./database');
+const database = require('../database');
 const fetch = require('node-fetch');
 const cheerio = require('cheerio');
 const pdfParse = require('pdf-parse');
-const embeddingSystem = require('./embedding-system');
-const { commandMap: musicCommandMap } = require('./src/commands/music');
-const CooldownManager = require('./src/core/cooldown-manager');
-const { recordCommandRun } = require('./src/utils/telemetry');
-const { commandFeatureMap, SLASH_EPHEMERAL_COMMANDS } = require('./src/core/command-registry');
-const { isFeatureGloballyEnabled, isFeatureEnabledForGuild } = require('./src/core/feature-flags');
-const memeCanvas = require('./src/utils/meme-canvas');
-const cryptoClient = require('./crypto-client');
-const vaultClient = require('./vault-client');
-const moderationFilters = require('./moderation-filters');
+const embeddingSystem = require('../embedding-system');
+const { commandMap: musicCommandMap } = require('../../commands/music');
+const CooldownManager = require('../../core/cooldown-manager');
+const { recordCommandRun } = require('../../utils/telemetry');
+const { commandFeatureMap, SLASH_EPHEMERAL_COMMANDS } = require('../../core/command-registry');
+const { isFeatureGloballyEnabled, isFeatureEnabledForGuild } = require('../../core/feature-flags');
+const memeCanvas = require('../../utils/meme-canvas');
+const cryptoClient = require('../crypto-client');
+const vaultClient = require('../vault-client');
+const moderationFilters = require('../moderation-filters');
 const NEWS_API_KEY = process.env.NEWS_API_KEY || null;
-const BrowserAgent = require('./src/agents/browserAgent');
-const AgentMonitor = require('./src/agents/agentMonitor');
-const AgentConfig = require('./src/agents/agentConfig');
-const RetryPolicy = require('./src/agents/retryPolicy');
-const AutoHealer = require('./src/agents/autoHealer');
-const CaptchaHandler = require('./src/agents/captchaHandler');
-const RobustnessEnhancer = require('./src/agents/robustnessEnhancer');
-const tempFiles = require('./src/utils/temp-files');
-const { sanitizePings: sanitizePingsUtil } = require('./src/utils/sanitize');
+const BrowserAgent = require('../../agents/browserAgent');
+const AgentMonitor = require('../../agents/agentMonitor');
+const AgentConfig = require('../../agents/agentConfig');
+const RetryPolicy = require('../../agents/retryPolicy');
+const AutoHealer = require('../../agents/autoHealer');
+const CaptchaHandler = require('../../agents/captchaHandler');
+const RobustnessEnhancer = require('../../agents/robustnessEnhancer');
+const tempFiles = require('../../utils/temp-files');
+const { sanitizePings: sanitizePingsUtil } = require('../../utils/sanitize');
 
 function isCommandEnabled(commandName) {
     const featureKey = commandFeatureMap.get(commandName);
