+++ tests/test-module-perf.js
