+++ tests/test-new-agent.js
