+++ tests/test-smart-tools.js
